move type to -node package
