+++ workspaces/explore/plugins/explore-node/src/index.ts
@@ -15,3 +15,4 @@
  */
 
 export { type ExploreToolProviderExtensionPoint, exploreToolProviderExtensionPoint } from './extensions';
+export type * from './types';
